+++ src/main/java/com/example/demo/model/entity/Deal.java
@@ -1,7 +1,6 @@
 package com.example.demo.model.entity;
 
 import lombok.Data;
-
 import javax.persistence.*;
 
 @Entity
@@ -10,12 +9,9 @@ import javax.persistence.*;
 public class Deal {
 
     @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @GeneratedValue(strategy=GenerationType.SEQUENCE,
-            generator="deal_id_seq")
-    @SequenceGenerator(name="deal_id_seq",
-            sequenceName="deal_id_seq", allocationSize=10)
-    private int id;
+    @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="deal_id_seq")
+    @SequenceGenerator(name="deal_id_seq", sequenceName="deal_id_seq", allocationSize=10)
+    private Long id;
     private int signed;
     private int value;
     private String currency;
